[ads] Add test::BuildRewardConfirmation and test::BuildNonRewardConfirmation utils
