+++ components/brave_ads/core/internal/account/confirmations/non_reward/non_reward_confirmation_unittest_util.h
@@ -1,0 +1,24 @@
+/* Copyright (c) 2024 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_CONFIRMATIONS_NON_REWARD_NON_REWARD_CONFIRMATION_UNITTEST_UTIL_H_
+#define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_CONFIRMATIONS_NON_REWARD_NON_REWARD_CONFIRMATION_UNITTEST_UTIL_H_
+
+#include <optional>
+
+namespace brave_ads {
+
+struct ConfirmationInfo;
+
+namespace test {
+
+std::optional<ConfirmationInfo> BuildNonRewardConfirmation(
+    bool should_use_random_uuids);
+
+}  // namespace test
+
+}  // namespace brave_ads
+
+#endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_CONFIRMATIONS_NON_REWARD_NON_REWARD_CONFIRMATION_UNITTEST_UTIL_H_
