+++ components/brave_ads/core/test/BUILD.gn
@@ -20,6 +20,8 @@ source_set("brave_ads_unit_tests") {
     "//brave/components/brave_ads/core/internal/account/confirmations/confirmations_unittest.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/confirmations_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/non_reward/non_reward_confirmation_util_unittest.cc",
+    "//brave/components/brave_ads/core/internal/account/confirmations/non_reward/non_reward_confirmation_unittest_util.cc",
+    "//brave/components/brave_ads/core/internal/account/confirmations/non_reward/non_reward_confirmation_unittest_util.h",
     "//brave/components/brave_ads/core/internal/account/confirmations/payload/confirmation_payload_json_writer_unittest.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/payload/reward_confirmation_payload_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/queue/confirmation_queue_delegate_mock.cc",
@@ -28,8 +30,8 @@ source_set("brave_ads_unit_tests") {
     "//brave/components/brave_ads/core/internal/account/confirmations/queue/confirmation_queue_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/reward/reward_confirmation_util_unittest.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/reward/reward_credential_json_writer_unittest.cc",
-    "//brave/components/brave_ads/core/internal/account/confirmations/reward/reward_unittest_util.cc",
-    "//brave/components/brave_ads/core/internal/account/confirmations/reward/reward_unittest_util.h",
+    "//brave/components/brave_ads/core/internal/account/confirmations/reward/reward_confirmation_unittest_util.cc",
+    "//brave/components/brave_ads/core/internal/account/confirmations/reward/reward_confirmation_unittest_util.h",
     "//brave/components/brave_ads/core/internal/account/confirmations/user_data_builder/confirmation_user_data_builder_unittest.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/user_data_builder/confirmation_user_data_builder_unittest_util.cc",
     "//brave/components/brave_ads/core/internal/account/confirmations/user_data_builder/confirmation_user_data_builder_unittest_util.h",
